+++ component/processor.go
@@ -178,47 +178,38 @@ type processorFactory struct {
 	CreateLogsProcessorFunc
 }
 
-// WithTracesProcessor overrides the default "error not supported" implementation for CreateTracesProcessor.
-// Deprecated: [v0.55.0] Use WithTracesProcessorAndStabilityLevel instead.
-func WithTracesProcessor(createTracesProcessor CreateTracesProcessorFunc) ProcessorFactoryOption {
-	return WithTracesProcessorAndStabilityLevel(createTracesProcessor, StabilityLevelUndefined)
-}
-
-// WithTracesProcessorAndStabilityLevel overrides the default "error not supported" implementation for CreateTracesProcessor and the default "undefined" stability level.
-func WithTracesProcessorAndStabilityLevel(createTracesProcessor CreateTracesProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
+// WithTracesProcessor overrides the default "error not supported" implementation for CreateTracesProcessor and the default "undefined" stability level.
+func WithTracesProcessor(createTracesProcessor CreateTracesProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
 	return processorFactoryOptionFunc(func(o *processorFactory) {
 		o.stability[config.TracesDataType] = sl
 		o.CreateTracesProcessorFunc = createTracesProcessor
 	})
 }
 
-// WithMetricsProcessor overrides the default "error not supported" implementation for CreateMetricsProcessor.
-// Deprecated: [v0.55.0] Use WithMetricsProcessorAndStabilityLevel instead.
-func WithMetricsProcessor(createMetricsProcessor CreateMetricsProcessorFunc) ProcessorFactoryOption {
-	return WithMetricsProcessorAndStabilityLevel(createMetricsProcessor, StabilityLevelUndefined)
-}
+// Deprecated: [v0.57.0] Use WithTracesProcessor instead.
+var WithTracesProcessorAndStabilityLevel = WithTracesProcessor
 
-// WithMetricsProcessorAndStabilityLevel overrides the default "error not supported" implementation for CreateMetricsProcessor and the default "undefined" stability level.
-func WithMetricsProcessorAndStabilityLevel(createMetricsProcessor CreateMetricsProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
+// WithMetricsProcessor overrides the default "error not supported" implementation for CreateMetricsProcessor and the default "undefined" stability level.
+func WithMetricsProcessor(createMetricsProcessor CreateMetricsProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
 	return processorFactoryOptionFunc(func(o *processorFactory) {
 		o.stability[config.MetricsDataType] = sl
 		o.CreateMetricsProcessorFunc = createMetricsProcessor
 	})
 }
 
-// WithLogsProcessor overrides the default "error not supported" implementation for CreateLogsProcessor.
-// Deprecated: [v0.55.0] Use WithLogsProcessorAndStabilityLevel instead.
-func WithLogsProcessor(createLogsProcessor CreateLogsProcessorFunc) ProcessorFactoryOption {
-	return WithLogsProcessorAndStabilityLevel(createLogsProcessor, StabilityLevelUndefined)
-}
+// Deprecated: [v0.57.0] Use WithMetricsProcessor instead.
+var WithMetricsProcessorAndStabilityLevel = WithMetricsProcessor
 
-// WithLogsProcessorAndStabilityLevel overrides the default "error not supported" implementation for CreateLogsProcessor and the default "undefined" stability level.
-func WithLogsProcessorAndStabilityLevel(createLogsProcessor CreateLogsProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
+// WithLogsProcessor overrides the default "error not supported" implementation for CreateLogsProcessor and the default "undefined" stability level.
+func WithLogsProcessor(createLogsProcessor CreateLogsProcessorFunc, sl StabilityLevel) ProcessorFactoryOption {
 	return processorFactoryOptionFunc(func(o *processorFactory) {
 		o.stability[config.LogsDataType] = sl
 		o.CreateLogsProcessorFunc = createLogsProcessor
 	})
 }
+
+// Deprecated: [v0.57.0] Use WithLogsProcessor instead.
+var WithLogsProcessorAndStabilityLevel = WithLogsProcessor
 
 // NewProcessorFactory returns a ProcessorFactory.
 func NewProcessorFactory(cfgType config.Type, createDefaultConfig ProcessorCreateDefaultConfigFunc, options ...ProcessorFactoryOption) ProcessorFactory {